Updated server.ts
Added options for deleting departments, roles, and employees, but one work. For the SELECT queries, results were changed to tables instead of a log.  Added a means of SELECTing from departments/roles/employees in order to provide a list instead of asking users to input, for example, the role  ID when selecting their role.