Got the deleteEmployee to work
This one was a doozy: added another pool.query into the start of startCli to always get a up-to-date list of employees. Added a global variable to store that current list. Implemented a findIndex() to find the exact employee in the array when one is chosen for deletion, extracted their ID, used THAT for the delete query, and then parameterized the query. There was the issue where the delete query did not work the first time, but was successful the second time. This led to investigating transactional locks, which was ultimately unhelpful. Will delete that and comments later.